Move eval-log-tree-builder.ts to query-evaluation-logging/eval-log-tree-builder.ts and related tests
